+++ ImageProcessing/cnn.py
@@ -1,0 +1,4 @@
+import hog
+
+def face_center(filename, model):
+    return hog.face_center(filename, model)
