+++ config.py
@@ -319,15 +319,6 @@ class Config:
             return self.conf.get("face", "locations_model")
         except:
             return "hog"
-    
-    def face_app_id(self) -> str:
-        return self.conf.get("face", "appid")
-
-    def face_api_key(self) -> str:
-        return self.conf.get("face", "key")
-
-    def face_app_secret(self) -> str:
-        return self.conf.get("face", "secret")
 
 
     @staticmethod
